Improve method lookup on embedded type
Previously, moxie assumed that only one type would be embedded within a
struct, and that the canonical receiver for that type would always be
the same as the name of the embedded type. This inaccuracy meant that
moxie produced non-compiling code under a few circumstances, including
when embedded types would themselves embed composed interfaces.

io.ReadWriter is used for testing here as a common example of a
composed interface, since it embeds both io.Reader and io.Writer.

This change switches to using types.LookupFieldOrMethod which should
be a much more accurate way of determining which embedded type to call
when falling through to the non-mocked implementations.